Extract upcoming episodes query into the libraries feature
